Final Animated Mode Fix Complete!
I continued tweaking with my formulas until I FINALLY got the proper behavior that I was looking for!  The AngleScroller1 object will dynamically adjust itself to bounce between the last two angles that the User provides once a cycle begins.  And on top of that, it ALWAYS takes the shortest distance between two angles!  HUZZAH!!!